refactor: extract assertOptions
