+++ packages/create-react-native-library/src/utils/assert.ts
@@ -1,5 +1,6 @@
 import kleur from 'kleur';
 import { spawn } from './spawn';
+import type { Answers, Question } from 'create-react-native-library';
 
 export async function assertNpx() {
   try {
@@ -16,6 +17,59 @@ export async function assertNpx() {
       process.exit(1);
     } else {
       throw error;
+    }
+  }
+}
+
+/**
+ * Makes sure the answers are in expected form and ends the process with error if they are not
+ */
+export function assertAnswers(questions: Question[], answers: Answers) {
+  for (const [key, value] of Object.entries(answers)) {
+    if (value == null) {
+      continue;
+    }
+
+    const question = questions.find((q) => q.name === key);
+
+    if (question == null) {
+      continue;
+    }
+
+    let valid = question.validate ? question.validate(String(value)) : true;
+
+    // We also need to guard against invalid choices
+    // If we don't already have a validation message to provide a better error
+    if (typeof valid !== 'string' && 'choices' in question) {
+      const choices =
+        typeof question.choices === 'function'
+          ? question.choices(
+              undefined,
+              // @ts-expect-error: it complains about optional values, but it should be fine
+              answers,
+              question
+            )
+          : question.choices;
+
+      if (choices && !choices.some((choice) => choice.value === value)) {
+        valid = `Supported values are - ${choices.map((c) =>
+          kleur.green(c.value)
+        )}`;
+      }
+    }
+
+    if (valid !== true) {
+      let message = `Invalid value ${kleur.red(
+        String(value)
+      )} passed for ${kleur.blue(key)}`;
+
+      if (typeof valid === 'string') {
+        message += `: ${valid}`;
+      }
+
+      console.log(message);
+
+      process.exit(1);
     }
   }
 }
